requirements.txt
django-ninja
psycopg2-binary
groq
python-dotenv